backend/requirements.txt
# Python dependencies
fastapi==0.110.1
uvicorn==0.27.1
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6
alembic==1.13.1
psycopg2-binary==2.9.9
pandas==2.2.0
pyarrow==15.0.0
httpx==0.26.0
python-dotenv==1.0.0
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
asyncpg==0.29.0
email-validator==2.1.0